requirements.txt
bs4
filelock
importlib_metadata>=4.2.0
jieba
numpy
packaging
regex
sacremoses
scikit-learn
scipy>=1.5.4
sentencepiece>=0.1.91
tensorboard
tokenizers>=0.10.1
datasets>=2.1.0
torch>=1.8.0
rouge
tqdm>=4.48.2
pillow
albumentations
einops
torchvision
